+++ src/state/sock_request_writing.c
@@ -28,7 +28,7 @@ unsigned sock_request_writing_write_handler(struct selector_key * key){
     buffer_read_adv(client_data->write_buffer, written_bytes);
     buffer_compact(client_data->write_buffer);
 
-    if(written_bytes < write_amount)
+    if(written_bytes < (ssize_t)write_amount)
         return SOCK_REQUEST_WRITING;
     else{
         print_connection_data(key);
